feat: update menu items when inserting or deleting

+++ backend/internal/handlers/review/service.go
@@ -71,6 +71,15 @@ func (s *Service) InsertReview(r ReviewDocument) (*ReviewDocument, error) {
 		return &r, err
 	}
 
+	// Append the review ID to the menu item's reviews array
+	if err := s.appendMenuItemReview(r); err != nil {
+		return &r, err
+	}
+	// Update menu item's average rating
+	if err := s.updateMenuItemAverageRatings(r); err != nil {
+		return &r, err
+	}
+
 	return &r, nil
 }
 
@@ -149,9 +158,31 @@ func (s *Service) UpdatePartialReview(id primitive.ObjectID, updated ReviewDocum
 func (s *Service) DeleteReview(id primitive.ObjectID) error {
 	ctx := context.Background()
 	filter := bson.M{"_id": id}
-
+	// First, get the review document to remove its ID from the menu item
+	deletedReview, errGet := s.GetReviewByID(id)
+	if errGet != nil {
+		return errGet
+	}
 	_, err := s.reviews.DeleteOne(ctx, filter)
-	return err
+	if err != nil {
+		return err
+	}
+
+	// Remove the review ID from the menu item's reviews array
+	if err := s.removeMenuItemReview(*deletedReview); err != nil {
+		return err
+	}
+
+	// Update the menu item's average rating
+	if err := s.updateMenuItemAverageRatings(*deletedReview); err != nil {
+		return err
+	}
+	// Update the restaurant's average rating
+	if err := s.updateRestaurantAverageRating(deletedReview.RestaurantID); err != nil {
+		return err
+	}
+	return nil
+
 }
 
 // CreateComment adds a new comment to a review
@@ -199,8 +230,24 @@ func (s *Service) GetComments(reviewID primitive.ObjectID) ([]CommentDocument, e
 	return comments, err
 }
 
-// appendMenuItemReviews appends the review ID to the menu item's reviews array
-func (s *Service) appendMenuItemReviews(reviewDoc ReviewDocument) error {
+func (s *Service) removeMenuItemReview(reviewDoc ReviewDocument) error {
+	ctx := context.Background()
+	menuItemID := reviewDoc.MenuItem
+	reviewID := reviewDoc.ID
+
+	// Update the menu item document to remove the review ID
+	update := bson.M{
+		"$pull": bson.M{
+			"reviews": reviewID,
+		},
+	}
+
+	_, err := s.menuItems.UpdateOne(ctx, bson.M{"_id": menuItemID}, update)
+	return err
+}
+
+// appendMenuItemReview appends the review ID to the menu item's reviews array
+func (s *Service) appendMenuItemReview(reviewDoc ReviewDocument) error {
 	ctx := context.Background()
 
 	menuId := reviewDoc.MenuItem
